Install Trusty HWE kernel
The version of the Ubuntu kernel we're currently using EOL'd on August 4th.
This means that whilst we might still receive updates for software installed
on the machine, we won't receive any security updates etc. for the kernel,
which is a bad thing.

To combat this, this commit installs the Trusty HWE kernel. Initially, it
installs it to a mirror to check this works properly when using Puppet
(it worked fine when using Apt in no-op mode).