Simplify HWE kernel installation
Whilst we could use an if statement to install the HWE kernel on only one
machine, removing the statement simplifies the deployment. To deploy to only
one machine, we can just deploy it to a single machine via the Fabric
deployment task.

We used to require the presence of `/usr/bin/hwe-support-status` and require
that it exited with 0 before ensuring that the presence of the
`linux-generic-lts-{distro}` and `linux-image-generic-lts-{distro}`
packages. This would be good as it would only install the packages when
absolutely required (as `hwe-support-status` checks whether they're required
by looking at the output of `uname -r` and doing some logic with it), but
since we need this on all machines eventually (our current kernel is EOL'd
already), we don't need this check.

Ensuring the presence of `update-manager-core` was required as it installs
`hwe-support-status`, as above. As we no longer use it, this can go too.