Replace `arnoudj/sudo` with `saz/sudo`
This commit replaces the `arnoudj/sudo` module with the `saz/sudo` module.
It turns out that we have lots of different ways of managing sudo, and it's
(probably) a Good Idea to stick to one, where possible.

`saz/sudo` is used by the main Puppet repository, so we're going to use it
here, too.